backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
pydantic-settings
pyjwt
passlib
python-multipart
aiofiles
httpx
